fix decimal handling for many queries

+++ libraries/kiwi/kiwi-sparql/src/main/java/org/apache/marmotta/kiwi/sparql/function/cast/NDecimalCast.java
@@ -62,7 +62,7 @@ public class NDecimalCast extends DecimalCast implements NativeFunction {
      */
     @Override
     public ValueType getReturnType() {
-        return ValueType.DOUBLE;
+        return ValueType.DECIMAL;
     }
 
     /**
